ai_chatbot/requirements.txt
openai>=1.0.0
pyttsx3>=2.90
SpeechRecognition>=3.8.1
Flask>=2.0.0
python-dotenv>=0.19.0